requirements.txt
tweepy==3.8.0
Pillow==7.0.0